refactor: Restructuring code
