+++ src/configs/defaultConfigs.ts
@@ -1,0 +1,29 @@
+export const defaultConfigs = {
+
+	/** Enable appearnace page */
+	'plugin.settings-app.appearance': false,
+
+	/** Enable theme section, on appearnace page */
+	'plugin.settings-app.appearance.theme': false,
+
+	/** Enable dark mode setting in theme section, on appearnace page */
+	'plugin.settings-app.appearance.theme.dark-mode': true,
+
+	/** Enable theme selection setting in theme section, on appearnace page */
+	'plugin.settings-app.appearance.theme.selection': false,
+
+	/** Enable appearnace page */
+	'plugin.settings-app.language': false,
+
+	/** Enable language section, on language page */
+	'plugin.settings-app.language.language': false,
+
+	/** Enable text direction setting in language section, on language page */
+	'plugin.settings-app.language.language.text-direction': true,
+
+	/** Enable language selection setting in language section, on language page */
+	'plugin.settings-app.language.language.selection': true,
+
+	/** Enable appearnace page */
+	'plugin.settings-app.about': false,
+};
