--- src/extras/ComponentList/__stories__/PluginList.stories.tsx
@@ -1,12 +1,0 @@
-// tslint:disable:no-console
-import { ComponentList } from '../ComponentList';
-import React from 'react';
-
-import storiesOf from '@bluebase/storybook-addon';
-
-storiesOf('Developers', module)
-
-.add('ComponentList', () => (
-	<ComponentList />
-))
-;
