+++ src/lang/index.ts
@@ -1,0 +1,5 @@
+import { ur } from './ur';
+
+export const lang = {
+	'bluebase.intl.messages.ur': ur,
+};
