+++ src/lang/index.web.ts
@@ -1,0 +1,3 @@
+export const lang = {
+	'bluebase.intl.messages.ur': import('./ur'),
+};
