+++ src/lang/ur.ts
@@ -1,0 +1,20 @@
+// tslint:disable: max-line-length
+import { IntlMessages } from '@bluebase/core';
+
+export const ur = (messages: IntlMessages) => ({
+	...messages,
+	'Auto': 'آٹو',
+	'Available Themes': 'دستیاب موضوعات',
+	'BlueBase Dark': 'بلیو بیز ڈارک',
+	'BlueBase Light': 'بلیو بیس لائٹ',
+	'Developed by': 'کی طرف سے تیار',
+	'Language': 'زبان',
+	'Left to Right': 'بائیں سے دائیں',
+	'Right to Left': 'دائیں سے بائیں',
+	'Text Direction': 'متن کی سمت',
+	'Text Direction will automatically changed based on selected language.': 'متن کی سمت منتخب کردہ زبان کی بنیاد پر خود بخود تبدیل ہوجائے گی.',
+	'Text will be displayed from Left to Right': 'متن بائیں سے دائیں سے دکھایا جائے گا',
+	'Text will be displayed from Right to Left': 'متن دائیں بائیں سے دکھایا جائے گا',
+	'Theme': 'خیالیہ',
+	'Version': 'ورژن',
+});
